muahahahhaha

+++ source/types/table/table_hash.c
@@ -6,7 +6,7 @@
 /*   By: bhildebr <bhildebr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/22 23:04:56 by bhildebr          #+#    #+#             */
-/*   Updated: 2024/02/23 02:10:29 by bhildebr         ###   ########.fr       */
+/*   Updated: 2024/02/23 23:47:47 by bhildebr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	table_hash(t_any key)
 	hash = 5381;
 	buffer = (unsigned char *)&key;
 	i = 0;
-	while (i < sizeof(t_any))
+	while (i < (int)sizeof(t_any))
 	{
 		hash = ((hash << 5) + hash) + buffer[i];
 	}
